refactor: merge gemma4_moe into gemma4 directory
Single directory matching the pattern of other model implementations.
Remove dead Gemma4MoeForCausalLM/Model/PreTrainedModel classes (unified
Gemma4ForCausalLM handles both via enable_moe_block). Clean up redundant
double-transpose in attention and fix histc float indices in for-loop
expert fallback. Add CHANGELOG entry for chat_template_kwargs.
