--- src/prime_rl/trainer/models/gemma4_moe/__init__.py
@@ -1,3 +1,0 @@
-from prime_rl.trainer.models.gemma4_moe.modeling_gemma4_moe import Gemma4MoeDecoderLayer
-
-__all__ = ["Gemma4MoeDecoderLayer"]
